requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
neo4j>=5.14.0
tabulate>=0.9.0